wiki-ingest-function/requirements.txt
azure-functions
azure-storage-blob>=12.28.0
tqdm
requests